% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToTable.R
\name{connectToTable}
\alias{connectToTable}
\title{Creates a connection to a table in the data base.}
\usage{
connectToTable(DBname, TableName)
}
\arguments{
\item{DBname}{}

\item{TableName}{}
}
\description{
This function will create a connection to the specified table and database.
}
