% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFact.R
\name{createFact}
\alias{createFact}
\title{Creates a Fact data table to be used as the main table in the power bi report.}
\usage{
createFact(DBname, fileLocation)
}
\arguments{
\item{DBname}{}

\item{fileLocation}{}
}
\description{
This function will create a table that will be called FACT in the power bi report.
}
