% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResponse.R
\name{createResponse}
\alias{createResponse}
\title{Creates a Response data table to be used as the response table in the power bi report.}
\usage{
createResponse(DBname, fileLocation)
}
\arguments{
\item{DBname}{}

\item{fileLocation}{}
}
\description{
This function will create a table that will be called Response in the power bi report.
}
