% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOneSheet.R
\name{saveOneSheet}
\alias{saveOneSheet}
\title{Creates a session data for Knect that will be saved in MS Excel as one sheet.}
\usage{
saveOneSheet(csvFileName, fileLocation)
}
\arguments{
\item{csvFileName}{}

\item{fileLocation}{}
}
\description{
This function will create a session data and will be saved in excel as one sheet. This should only be used by onsite managers doing a report for Knect.
}
